feat(bench): extract Claude JSON parser + LLMMeter accumulator
Phase 2.2: bench/metrics.py centralizes parsing of `claude --print
--output-format json`. parse_claude_json normalizes to {final_answer,
tokens_in (billable only), tokens_out, dollars, is_error, subtype,
num_turns}. LLMMeter dataclass accumulates across multiple Claude calls
(used by Phase 4 loop variants). claude_plain refactors to use it.
Parser tests migrate to test_bench_metrics.py.
